// src/lib/api/generated/model/scheduleSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface ScheduleSchema {
  confirmed_date: string;
  confirmed_end_time: string;
  confirmed_start_time: string;
  internal_notes?: string;
}